+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 
 # Changelog
 
+## v0.4.0
+
+- Project renamed
+- Improved non-aliased syntax
+- Added `input` and `textarea` extensions
+- One extension per node
+- Fixed several bugs
+- Added more tests
+- Added demos
+- Added documentation
+
 ## v0.3.0
 
 - Stable api